requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
affine=2.3.1=pyhd8ed1ab_0
asttokens=2.0.5=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
black=22.10.0=pypi_0
blosc=1.21.1=h74325e0_3
bmipy=2.0=pypi_0
boost-cpp=1.78.0=h9f4b32c_1
branca=0.5.0=pyhd8ed1ab_0
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=he774522_0
ca-certificates=2022.10.11=haa95532_0
cairo=1.16.0=hd694305_1014
cartopy=0.21.0=py310h05326cb_2
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h628cb3f_2
cfitsio=4.1.0=h5a969a9_0
cftime=1.6.2=pypi_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=win_pyhd8ed1ab_2
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
colorama=0.4.6=pypi_0
contextily=1.2.0=pyhd8ed1ab_0
contourpy=1.0.6=py310h232114e_0
country-bounding-boxes=0.2.3=pypi_0
cryptography=38.0.3=py310h6e82f81_0
curl=7.85.0=h2bbff1b_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6c2663c_0
fiona=1.8.22=py310h30f9a14_0
folium=0.13.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.0=h720f74d_1
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.38.0=py310h8d17308_1
freetype=2.12.1=ha860e81_0
freexl=1.0.6=h2bbff1b_0
gdal=3.5.2=py310h644bc08_4
geographiclib=1.52=pyhd8ed1ab_0
geopandas=0.12.1=pyhd8ed1ab_0
geopandas-base=0.12.1=pyha770c72_0
geoplot=0.5.1=pyhd8ed1ab_0
geopy=2.3.0=pyhd8ed1ab_0
geos=3.11.0=h39d44d4_0
geotiff=1.7.1=h714bc5f_3
gettext=0.21.1=h5728263_0
hdf4=4.2.15=h0e5069d_4
hdf5=1.12.2=nompi_h57737ce_100
icu=70.1=h0e60522_0
idna=3.4=pyhd8ed1ab_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2022.1.0=h59b6b97_3788
ipykernel=6.15.2=py310haa95532_0
ipython=8.4.0=py310haa95532_0
iso3166=2.1.1=pypi_0
jedi=0.18.1=py310haa95532_1
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h2bbff1b_0
jupyter_client=7.3.5=py310haa95532_0
jupyter_core=4.11.1=py310haa95532_0
kealib=1.4.15=hdf81f3a_1
kiwisolver=1.4.4=py310h232114e_1
krb5=1.19.3=hc8ab02b_0
lcms2=2.12=h83e58a3_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=16_win64_mkl
libcurl=7.85.0=h86230a5_0
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=hd77b12b_4
libgdal=3.5.2=h2477998_4
libglib=2.74.0=h79619a9_0
libiconv=1.17=h8ffe710_0
libkml=1.3.0=hf2ab4e4_1015
liblapack=3.9.0=16_win64_mkl
libnetcdf=4.8.1=nompi_h85765be_104
libpng=1.6.38=h19919ed_0
libpq=14.5=h1ea2d34_0
librttopo=1.1.0=h2842628_11
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h39d44d4_4
libspatialite=5.0.1=ha17912d_18
libsqlite=3.39.4=hcfcfb64_0
libssh2=1.10.0=h9a1e1f7_3
libtiff=4.4.0=h8a3f274_1
libwebp-base=1.2.4=h2bbff1b_0
libxcb=1.13=hcd874cb_1004
libxml2=2.9.14=h0ad7f3c_0
libzip=1.9.2=h519de47_1
libzlib=1.2.13=hcfcfb64_4
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mapclassify=2.4.3=pyhd8ed1ab_0
markupsafe=2.1.1=py310h8d17308_2
matplotlib-base=3.6.2=py310h51140c5_0
matplotlib-inline=0.1.6=py310haa95532_0
mercantile=1.2.1=pyhd8ed1ab_0
mkl=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
mypy-extensions=0.4.3=pypi_0
nest-asyncio=1.5.5=py310haa95532_0
netcdf4=1.6.1=pypi_0
networkx=2.8.8=pyhd8ed1ab_0
numpy=1.23.4=py310h4a8f9c9_0
openjpeg=2.5.0=hc9384bd_1
openssl=3.0.7=hcfcfb64_0
owslib=0.27.2=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.1=py310h1c4a608_1
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.10.1=pypi_0
pcre=8.45=hd77b12b_0
pcre2=10.37=h0ff8eda_1
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hd4fb230_3
pip=22.3=pyhd8ed1ab_0
pixman=0.40.0=h2bbff1b_1
platformdirs=2.5.2=pypi_0
pluggy=1.0.0=py310haa95532_1
poppler=22.10.0=ha6c1112_0
poppler-data=0.4.11=hd8ed1ab_0
postgresql=14.5=he353ca9_0
proj=9.0.1=h1cfcee9_1
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.0=py310h2bbff1b_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=py310haa95532_0
pyproj=3.4.0=py310h6377384_0
pyshp=2.3.1=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
pytest=7.1.2=py310haa95532_0
python=3.10.6=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2022.6=pyhd8ed1ab_0
pywin32=302=py310h2bbff1b_2
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310hd77b12b_0
rasterio=1.3.3=pypi_0
requests=2.28.1=pyhd8ed1ab_1
rioxarray=0.12.4=pypi_0
rtree=1.0.1=py310h1cbd46b_1
scikit-learn=1.1.3=py310had3394f_1
scipy=1.9.3=py310h578b7cb_1
seaborn-base=0.12.1=pyhd8ed1ab_0
setuptools=65.5.0=py310haa95532_0
shapely=1.8.5=py310h3734685_1
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
snuggs=1.4.7=py_0
soilgrids=0.1.3=pypi_0
sqlite=3.39.4=hcfcfb64_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.6.0=h59b6b97_0
threadpoolctl=3.1.0=pyh8a188c0_0
tiledb=2.11.3=h3132609_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=py310haa95532_0
tornado=6.2=py310h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022f=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
unicodedata2=15.0.0=py310h8d17308_0
urllib3=1.26.12=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.29.30139=h890b9b1_8
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wincertstore=0.2=py310haa95532_2
xarray=2022.10.0=pypi_0
xerces-c=3.2.4=h63175ca_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xyzservices=2022.9.0=pyhd8ed1ab_0
xz=5.2.6=h8cc25b3_0
zeromq=4.3.4=hd77b12b_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h19a0ad4_0
